brochure/requirements.txt
python-dotenv
ipython
requests
beautifulsoup4
openai